requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
plotly==5.18.0
pytest==7.4.3
pytest-mock==3.12.0
